fix authstate and login
