+++ app/index.html
@@ -15,7 +15,6 @@
 
     <link rel="shortcut icon" type="image/png" href="/assets/icon/favicon.ico" />
 
-    <!-- add to homescreen for ios -->
     <meta name="apple-mobile-web-app-capable" content="yes" />
     <meta name="apple-mobile-web-app-title" content="BiodiversityHub BC" />
     <meta name="apple-mobile-web-app-status-bar-style" content="black" />
