+++ app/src/config/config.interface.ts
@@ -1,0 +1,14 @@
+export interface IConfig {
+  API_HOST: string;
+  CHANGE_VERSION: string;
+  NODE_ENV: string;
+  VERSION: string;
+  KEYCLOAK_CONFIG: {
+    authority: string;
+    realm: string;
+    clientId: string;
+  };
+  SITEMINDER_LOGOUT_URL: string;
+  MAX_UPLOAD_NUM_FILES: number;
+  MAX_UPLOAD_FILE_SIZE: number;
+}
